Add Namespace by reusing Jupiter's Namespace class
Issue: #2816
