+++ junit-platform-launcher/src/main/java/org/junit/platform/launcher/LauncherSession.java
@@ -11,9 +11,9 @@
 package org.junit.platform.launcher;
 
 import static org.apiguardian.api.API.Status.STABLE;
-import static org.junit.platform.engine.support.store.ResourceContext.*;
 
 import org.apiguardian.api.API;
+import org.junit.platform.engine.Namespace;
 import org.junit.platform.engine.support.store.NamespacedHierarchicalStore;
 import org.junit.platform.launcher.core.LauncherFactory;
